Added Application#processorConfig to the saved config
Added field Application.APP_VERSION
Added Application#processorConfig to the saved config
Added method Application#getProcessorConfig

Added methods: StringUtils.parseLong, StringUtils.compareVersions
Now App's config is reset to defaults if it was made on a different version of the app

Added method ArgumentsParser#isSpecified
Now only loading specified command line arguments

Made method KeyValueData#getOrDefault automatically convert Numbers to others
Made method KeyValueParser.parseKeyValuePairs only store numbers as either Doubles or Longs

Added method ConfigManager.stopEvent, which stops event propagation

Added static fields to ProcessorConfig which are used to get the max, min and defaults for each property
Made all primitive fields on ProcessorConfig private and added methods to set and get them

+++ src/main/java/io/github/hds/pemu/utils/KeyValueParser.java
@@ -60,11 +60,11 @@ public class KeyValueParser {
         if (number == null) return null;
 
         try {
-            return StringUtils.parseInt(number);
+            return StringUtils.parseLong(number);
         } catch (Exception ignored) { }
 
         try {
-            return Float.parseFloat(number);
+            return Double.parseDouble(number);
         } catch (Exception ignored) { }
 
         return null;
